requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
jinja2==3.1.5
aiofiles==24.1.0
google-api-python-client==2.164.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
anthropic==0.52.0
apscheduler==3.10.4
python-dotenv==1.0.1
